requirements.txt
anthropic>=0.40.0
slack-sdk>=3.27.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
google-auth>=2.25.0
google-auth-oauthlib>=1.2.0
google-api-python-client>=2.110.0
pydantic>=2.5.0
cryptography>=41.0.0
httpx>=0.25.0
google-genai>=1.0.0
